Firmware: use the generic type 'int' to describe colors internally

+++ firmware/apps/color/include/color.h
@@ -19,7 +19,6 @@
 
 extern int color_init(void);
 
-extern int color_read_data(uint16_t *r, uint16_t *g, uint16_t *b,
-                           uint16_t *clear);
+extern int color_read_data(int *r, int *g, int *b, int *clear);
 
 extern int color_set_led_usage(int val);
